// src/app/models/employees.ts
export interface EmployeesModel {
    idEmployee: number;
    name: string;
    business: string;
    branch: string;
    job: string;
    department: string;
    antiquity: string;
    laborCompliance: string;
    employmentFile: string;
    actions: string;
    avatar: string;
    status: string;
    color: string;
}